requirements.txt
numpy==1.21.0
typing-extensions==4.3.0
scipy==1.10.0
scikit-learn==1.3.2
scikit-image==0.19.3
packaging==24.0
ml-collections==0.1.0
tensorflow-gan==2.0.0
tensorflow_io==0.34.0
tensorflow_datasets==3.1.0
tensorflow-gpu==2.8.0
tensorflow-probability==0.16.0
tensorflow-addons==0.12.0
tensorboard==2.8.0
absl-py==0.10.0
torch==2.1.2
torchmetrics==1.3.2
pytorch-ignite==0.5.0.post2
torchvision==0.16.2
matplotlib==3.7.5
natsort==8.4.0
einops==0.7.0
kornia==0.7.2
opencv-python==4.9.0.80
lpips==0.1.4
# ninja==1.11.1.1

# conda create -n sde2 python=3.8
# pip install -r requirements.txt
# export CUDA_HOME=/home/dsense/miniconda3/envs/sde2
# Try changing line 956 of /home/dsense/miniconda3/envs/sde2/lib/python3.9/site-packages/torch/utils/cpp_extension.py
#   from ['ninja', '-v'] to ['ninja', '--version']


# Also execute:
# conda install nvidia::cuda-nvcc
# wget https://github.com/ninja-build/ninja/releases/download/v1.8.2/ninja-linux.zip
# sudo unzip ninja-linux.zip -d /usr/local/bin/
# sudo update-alternatives --install /usr/bin/ninja ninja /usr/local/bin/ninja 1 --force
# export CUDA_HOME=/home/dsense/miniconda3/envs/sde2